Add deeplinking functionality to the push notifications.
When the app is opened through a push notification, it will handle it as
follows:

If the app is in the background or closed when the push notification
arrives, it will check whether the push notification contains a URL. If
it doesn't, the homepage will simply be shown. If it does have a URL, it
checks whether the URL is recognised as an URL for which we have a
deeplinking action. If so, it will open the corresponding screen
in the app. Otherwise, it will open the URL in the user's browser.

If the app is in the foreground when the notification arrives, it will
open an alert with the notifications title and contents. If the
notification has an URL, we show two buttons: one for opening the URL,
one for dismissing the notification. Otherwise, there is only a button
for dismissing the notification.